requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0

# OpenAI for API calls
openai==1.3.0
supabase==2.0.2

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# Lightweight alternatives
# numpy==1.26.0  # Removed - too heavy for Vercel
# scipy==1.11.4  # Removed - too heavy for Vercel
# chromadb==0.4.18  # Removed - too heavy for Vercel
# llama-index==0.9.48  # Removed - too heavy for Vercel
# langchain==0.1.0  # Removed - too heavy for Vercel
